Find ALL files named *.old, not just *.php.old (#3091)
There are other files in the vendor folder, they are not only .php files there, and sometimes these files need patching too. For example *.js or *.css files. Since the behavior of only finding *.php.old files is not mentioned in the documentation, nor implied in the package's name ("vendor-patches"), nor particularly useful for anything else than a speed optimization (and probably not even that), I removed this artificial limitation.